Updated Controller to use int instead of long

+++ msmpm/Controllers/PropertiesController.cs
@@ -29,7 +29,7 @@ namespace MSMBackend.Controllers
 
         // GET: api/Properties/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Property>> GetProperty(long id)
+        public async Task<ActionResult<Property>> GetProperty(int id)
         {
             var @property = await _context.Properties.FindAsync(id);
 
@@ -45,7 +45,7 @@ namespace MSMBackend.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutProperty(long id, Property @property)
+        public async Task<IActionResult> PutProperty(int id, Property @property)
         {
             if (id != @property.Id)
             {
@@ -87,7 +87,7 @@ namespace MSMBackend.Controllers
 
         // DELETE: api/Properties/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Property>> DeleteProperty(long id)
+        public async Task<ActionResult<Property>> DeleteProperty(int id)
         {
             var @property = await _context.Properties.FindAsync(id);
             if (@property == null)
@@ -101,7 +101,7 @@ namespace MSMBackend.Controllers
             return @property;
         }
 
-        private bool PropertyExists(long id)
+        private bool PropertyExists(int id)
         {
             return _context.Properties.Any(e => e.Id == id);
         }
